Fix handling of node-format when output-format-type is set
Current behavior is that if output-format-type is set node-format is
ignored. This includes the case where output-format-type is set in a
config file, so if your config file sets a default format it's
impossible to override it on the command line.

One solution is to avoid setting a default output format in config
files. While that works, it's worth noting unidic-mecab's default config
file includes a default output format. If someone tries to use
node-format without being aware of this they might just think it doesn't
work.

This change makes it so that if node-format is explicitly set
output-format-type is ignored. If output-format-type isn't used behavior
is unchanged; this only changes things when output-format-type and
node-format are both set. -POLM